Fix method resolution to warn about things that don't exist on the declared type.
Not sure how/when this got broken, but we were only warning if no name in the whole world matched. The fix for this is in "value.dart"

Also cleaned up a lot of the compiler code to keep it warning free. There are some warnings in leg, but I left those as I'm not sure what should be done.

Review URL: http://codereview.chromium.org//8497064

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@1383 260f80e4-7a28-3924-810f-c04153c831b5