Partial fix for Issue 709.
Method dispatch on a hidden native class goes according to the native
name, not the Dart name (which can be different).

There is still an issue with the hidden name - frog needs to avoid
using it because it might not be inaccessible on all targets.

BUG=
TEST=

Review URL: http://codereview.chromium.org//8974005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@2511 260f80e4-7a28-3924-810f-c04153c831b5