Disable dart:html tests broken by WebKit changes.
I hate to do this but it might take a little while to fix and it will be helpful to have client green in the meantime.

Issue 1005 needs to be fixed before these can be re-enabled.
The initXXXEvent methods which these tests use have been removed from WebKit.  A workable solution needs to be added to the dart:dom layer.

BUG=
TEST=

Review URL: http://codereview.chromium.org//9015016

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@2883 260f80e4-7a28-3924-810f-c04153c831b5