Introducing the File Explorer view.
Caveat emptor: There are a number of bits that need further investigation.  Most notably shimming in an alternative implementation of org.eclipse.core.internal.resources.Project (and friends) so we can avoid dirtying projects with .project file litter (to the chagrin of SCMs everywhere).  There's also hardening we want to do to the wizard to keep users from falling off the rails (attempting to create projects with silly names or nesting them, etc.).  But the core is here.  In particular, we've got:

   * project creation and deletion
   * resource rename, delete and move
   * view link with editor

The view is deliberately file-centric with the exception of special nodes for referenced libraries.  Alternative mixed representations can be explored in future CLs.

To reflect the experimental nature of this view, it is hidden behind the bleeding edge flag.  To enable it one needs to set the com.google.dart.tools.core/bleedingEdge option to true.
Review URL: http://codereview.chromium.org//9151019

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@3163 260f80e4-7a28-3924-810f-c04153c831b5