Refactor DOM generator.
Each target (approximately) is now represented by a 'System'.  This
should be helpful in merging in the dart:hmtl code - it can be another
System (or 2 - one for frog, one for VM).

There are more changes to come.  In particular, I want to make the
class/interface templates per-system.

Checkpoint

BUG=
TEST=

Review URL: http://codereview.chromium.org//9181010

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@3172 260f80e4-7a28-3924-810f-c04153c831b5