This change gets us set up for type inference in frog. Essentially: unification of Values, computing fixed point in loops, and tracking the static type separately from the inferred type. This also fixes the old TODO about how we should not be doing inference on "final". I backed out of the more aggressive (cross method Value propegation) changes to simplify things the initial checkin, but I'll look at them after this is in.
A few places had to be fixed because they were relying on members that are on subtypes. I'm not sure why this was working before--but likely the Value.staticType cleanup fixed it incidentally.

Note: Inference is off by default (enabled with --infer_types) so I can take a bit more time to look at it, there's a few test failures still. But the early stats look pretty good. Somewhere around half of dynamic calls/type asserts in Frog go away. (it gets worse in the minfrog in this CR, but that's because we aren't cheating on "final" anymore)

Review URL: http://codereview.chromium.org//9186061

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@3277 260f80e4-7a28-3924-810f-c04153c831b5