This CL generates factory provider classes for any interface with a [Constructor] or [NamedConstructor] extended attribute.
The factory provider is generated from a template, using the pattern of an IDL interface specific template for the exceptional cases and a general template for the other cases.  I used the special case to adjust the number of arguments in the JavaScript call - longer term I think we might want to be a bit more intelligent about default argument values.

This CL generates for frog and dummy, the others need to be added.  Once the others have been added we might see an opportunity for abstracting out some of the mechanism but I it might turn out they are all slightly different.

BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com//9428050

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@4560 260f80e4-7a28-3924-810f-c04153c831b5