This CL brings in changes from http://codereview.chromium.org/9290033/ since the navigator UX has changed in favor of the Eclipse project story.
Updates to the Dart core to update the set of top-level libraries in each Dart project for file adds, removes and changes. The consequence of this is that when a user adds, removes or changes a dart file, the correct content will be compiled.
Review URL: https://chromiumcodereview.appspot.com//9555012

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@4807 260f80e4-7a28-3924-810f-c04153c831b5