Fix for highlighting issue (http://code.google.com/p/dart/issues/detail?id=2550).
I left the code commented out because I seem to remember needing to add it for some reason. I can't reproduce a problem with it commented out (and it fixes this particular issue), but I didn't want to loose it in case we re-discover the problem.
Review URL: https://chromiumcodereview.appspot.com//10038029

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@6476 260f80e4-7a28-3924-810f-c04153c831b5